requirements.txt
pypdf>=4.0.0
ragas>=0.1.7
openai>=1.10.0
python-dotenv>=1.0.0
pandas>=2.1.0
tqdm>=4.66.0
jinja2>=3.1.0
<<<<<<< HEAD
transformers>=4.40.0
torch
sentence-transformers
faiss-cpu
accelerate
=======
chromadb>=0.5.0
sentence-transformers>=2.7.0
>>>>>>> origin/nxhai
